+++ Sstv.DomainExceptions/ErrorDescriptionAttribute.cs
@@ -17,6 +17,11 @@ public sealed class ErrorDescriptionAttribute : Attribute
     public string? Description { get; set; }
 
     /// <summary>
+    /// Criticality level.
+    /// </summary>
+    public Level Level { get; set; }
+
+    /// <summary>
     /// Uri for help link.
     /// </summary>
     public string? HelpLink { get; set; }