+++ Sstv.Host/ErrorCodes.cs
@@ -2,32 +2,33 @@
 
 namespace Sstv.Host;
 
-[ErrorDescription(Prefix = "SSTV", HelpLink = "https://help.myproject.ru/error-codes/{0}")]
-[ExceptionConfig(ClassName = "CoreException")]
+[ErrorDescription(Prefix = "SSTV", HelpLink = "https://help.myproject.ru/error-codes/{0}", Level = Level.Critical)]
+[ExceptionConfig(ClassName = "FirstException")]
 public enum ErrorCodes
 {
     [ErrorDescription(
         Description = "Unhandled error code",
-        HelpLink = "https://help.myproject.ru/error-codes/nothing-here")]
+        HelpLink = "https://help.myproject.ru/error-codes/nothing-here",
+        Level = Level.Fatal)]
     Default = 0,
+
+    [ErrorDescription(Description = "Invalid data", Level = Level.Critical)]
+    InvalidData = 1,
 
     [ErrorDescription(
         Description = "You have not enough money",
-        HelpLink = "https://help.myproject.ru/error-codes/not-enough-money")]
+        HelpLink = "https://help.myproject.ru/error-codes/not-enough-money",
+        Level = Level.NotError)]
     NotEnoughMoney = 10001,
 
     [ErrorDescription(Prefix = "DIF", Description = "Another prefix example")]
     SomethingBadHappen = 10002,
 
-    [Obsolete("Don't use this error code because it obsolete :)")]
-    [ErrorDescription(Prefix = "DIF", Description = "Obsolete error code in enum")]
-    ObsoleteErrorCode = 10003,
-
     [ErrorDescription(
         Description = "Help link with template in enum member attribute",
         HelpLink = "https://help.myproject.ru/{0}/error-code"
     )]
-    WhateverElse = 10004,
+    WhateverElse = 10003,
 }
 
 [ErrorDescription(Prefix = "SDPV", HelpLink = "https://help.second-myproject.ru/error-codes/{0}")]