+++ Sstv.Host/Program.cs
@@ -2,13 +2,11 @@
 // https://github.com/dotnet/roslyn-analyzers/issues/6141
 #pragma warning disable CA1852
 
-using System.Text.Json;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Resources;
-using Sstv.DomainExceptions;
 using Sstv.DomainExceptions.Extensions.DependencyInjection;
-using Sstv.DomainExceptions.Extensions.ProblemDetails;
 using Sstv.Host;
+using System.Text.Json;
 using System.Text.Json.Serialization;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -22,48 +20,12 @@ builder.Services.AddOpenTelemetry()
         o.AddPrometheusExporter();
     });
 
-builder.Services.AddDomainExceptions(b =>
-{
-    b.WithErrorCodesDescriptionSource(CoreException.ErrorCodesDescriptionSource);
-    b.WithErrorCodesDescriptionSource(SecondErrorCodesException.ErrorCodesDescriptionSource);
-    b.WithErrorCodesDescriptionFromConfiguration();
-    b.UseDomainExceptionHandler();
-
-    b.ConfigureSettings = (sp, settings) =>
-    {
-        settings.GenerateExceptionIdAutomatically = true;     // default value
-        settings.CollectErrorCodesMetricAutomatically = true; // default value
-        settings.ThrowIfHasNoErrorCodeDescription = true;     // default value
-        settings.ErrorCodesDescriptionSource = null;          // manually set your own error description source instance
-        settings.DefaultErrorDescriptionProvider =            // override default error description func
-            errorCode => new ErrorDescription(errorCode, "N/A"); // default func
-
-        settings.OnExceptionCreated += exception =>
-        {
-            Console.WriteLine(exception.ToString());
-        };
-    };
-});
+builder.Services.AddDomainException();
+builder.Services.AddProblemDetail();
 
 builder.Services
-    .AddProblemDetails(x =>
-    {
-        x.CustomizeProblemDetails = context =>
-        {
-            if (context.Exception is DomainException de)
-            {
-                context.ProblemDetails.Status = context.HttpContext.Response.StatusCode = ErrorCodeMapping.MapToStatusCode(de.ErrorCode);
-            }
-            else
-            {
-                context.ProblemDetails = new ErrorCodeProblemDetails(ErrorCodes.Default.GetDescription())
-                {
-                    Status = context.HttpContext.Response.StatusCode = StatusCodes.Status500InternalServerError
-                };
-            }
-        };
-    })
     .AddControllers()
+    .AddValidationProblemDetails()
     .AddJsonOptions(options =>
     {
         options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());