+++ Sstv.DomainExceptions.Extensions.ProblemDetails/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.0] - 2024-12-06
+
+- No changes here
+
 ## [2.2.0] - 2024-02-18
 
 ### Added