+++ Sstv.DomainExceptions/CHANGELOG.md
@@ -14,12 +14,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.0] - 2024-12-06
+
+- CriticalityLevel added to [ErrorDescription], also added to metric label, logs, api response. So you can use it to differ codes by its error criticality level.
+
+BREAKING CHANGES: 
+- Dropped field "IsObsolete" on [ErrorDescription]
+- DomainExceptionSettings.OnExceptionCreated replaced by DomainExceptionSettings.OnErrorCreated callback
+
 ## [2.1.0] - 2024-02-12
 
 - Adds [ExceptionConfig] attribute, for configuring the name of generated exception.
 - Exception now generated as partial class
 - AsException extensions method renamed to ToException
-
 
 ## [2.0.0] - 2024-02-11